fix class names and minor changes
