+++ modules/integration/test-ui-sp-editor/findbugs-exclude.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  ~ Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~ Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
   ~
   ~ WSO2 Inc. licenses this file to you under the Apache License,
   ~ Version 2.0 (the "License"); you may not use this file except
@@ -19,7 +19,5 @@
 
 <FindBugsFilter>
     <Match>
-        <!--Class name="org.wso2.sp.test"/>
-        <Bug pattern="Some input files use unchecked or unsafe operations."/-->
     </Match>
 </FindBugsFilter>
