+++ modules/integration/test-ui-sp-editor/src/main/java/org/wso2/sp/selenium/components/pageObjects/headerContainerObjects/HeaderDropDownDialogsObject.java
@@ -16,41 +16,33 @@
 
 package org.wso2.sp.selenium.components.pageObjects.headerContainerObjects;
 
+
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-public class toolBarContainerObject {
+public class HeaderDropDownDialogsObject {
     private WebDriver driver;
-    private By run = By.className("run_btn");
-    private By debug = By.className("debug_btn");
-    private By stop = By.className("stop_btn");
-    private By revert = By.className("revert_btn");
 
-    public toolBarContainerObject(WebDriver driver) {
+    public HeaderDropDownDialogsObject(WebDriver driver) {
         this.driver = driver;
     }
 
-    public void clickRun() {
-        driver.findElement(run).click();
+    public void clickSamples(String sampleID) {
+        driver.findElement(By.id(sampleID)).click();
     }
 
-    public void clickDebug() {
-        driver.findElement(debug).click();
+    public void selectFromDropdown(String dropdownID, String valueToBeSelected) {
+        driver.findElement(By.id(dropdownID)).click();
+        driver.findElement(By.cssSelector("option[value=" + valueToBeSelected + "]")).click();
     }
 
-    public void clickStop() {
-        driver.findElement(stop).click();
-    }
-
-    public void clickRevert() {
-        driver.findElement(revert).click();
-    }
-
-    public void waitForElement(int seconds, String waitConditionLocator) {
-        WebDriverWait wait = new WebDriverWait(driver, seconds);
-        wait.until(ExpectedConditions.visibilityOfElementLocated(By.xpath(waitConditionLocator)));
+    public void clickGuideClose() throws InterruptedException {
+        WebDriverWait wait = new WebDriverWait(driver, 30);
+        wait.until(ExpectedConditions.presenceOfElementLocated(By.xpath("//*[@id=\"guideDialog\"]/div/div/div[1]/button")));
+        driver.findElement(By.xpath("//*[@id=\"guideDialog\"]/div/div/div[1]/button")).click();
+        Thread.sleep(500);
     }
 
 }
