+++ modules/integration/test-ui-sp-editor/src/test/resources/testng.xml
@@ -1,5 +1,5 @@
 <!--
-  ~ Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~ Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
   ~
   ~ WSO2 Inc. licenses this file to you under the Apache License,
   ~ Version 2.0 (the "License"); you may not use this file except
@@ -18,9 +18,9 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 <suite name="suite of suites">
     <test name="testcase">
-<classes>
-    <class name="org.wso2.sp.selenium.spTestCases.testCases"/>
-</classes>
+        <classes>
+            <class name="org.wso2.sp.selenium.spTestCases.testCases"/>
+        </classes>
     </test>
 </suite>
